Added AcademicResultsDB Back to the project. Azza if you are reading this you will not be to build till you install Visual Studio. Also implemented Akode Importer for 2nd Sem and Board Results

+++ DEARS/DataAccess/MarksExamCW.vb
@@ -30,10 +30,10 @@ Partial Public Class MarksExamCW
     Public Property SemesterId As Integer
     Public Property CourseId As Integer
     Public Property StudentId As Integer
-    Public Property CWMark As Nullable(Of Decimal)
+    Public Property CWMark As Nullable(Of Double)
     Public Property Present As Nullable(Of Boolean)
     Public Property Excuse As Nullable(Of Boolean)
-    Public Property ExamMark As Nullable(Of Decimal)
+    Public Property ExamMark As Nullable(Of Double)
 
     Public Overridable Property CourseEnrollment As CourseEnrollment
     Public Overridable Property Course As Course
